Autoversioning commit:  a non-deltaV client made a change to /1/icodrops.php

+++ icodrops.php
@@ -15,11 +15,13 @@ foreach($str2 as $k=>$v){
     $contents1=file_get_contents_https($url);
     $arr[$i]['githhuburl']=$githuburl="https://github.com/".explode("\"",explode("https://github.com/",$contents1)[1])[0];
     $i++;
+    if($arr[$i]['name']!=''){
     $ret = $kv->delete('icodropsproducts:'.$i);
         $kv->add('icodropsproducts:'.$i, json_encode($arr[$i],true));
     $sql='insert into ico_Analysis (name,Github_url,DataSource) values("'.$arr[$i]['name'].'","'.$arr[$i]['githuburl'].'","icodrops");';
     MySQLRunSQL($sql);
      $kv->get('icodropsproducts:'.$i);
+    }
 }
 
 $url = 'https://icodrops.com/category/upcoming-ico/';
@@ -33,11 +35,13 @@ foreach($str2 as $k=>$v){
     $contents1=file_get_contents_https($url);
     $arr[$i]['githhuburl']=$githuburl="https://github.com/".explode("\"",explode("https://github.com/",$contents1)[1])[0];
     $i++;
+    if($arr[$i]['name']!=''){
     $ret = $kv->delete('icodropsproducts:'.$i);
         $kv->add('icodropsproducts:'.$i, json_encode($arr[$i],true));
     $sql='insert into ico_Analysis (name,Github_url,DataSource) values("'.$arr[$i]['name'].'","'.$arr[$i]['githuburl'].'","icodrops");';
     MySQLRunSQL($sql);
      $kv->get('icodropsproducts:'.$i);
+    }
 }
 $i=0;
 $url = 'https://icodrops.com/category/ended-ico/';
@@ -51,11 +55,13 @@ foreach($str2 as $k=>$v){
     $contents1=file_get_contents_https($url);
     $arr[$i]['githhuburl']=$githuburl="https://github.com/".explode("\"",explode("https://github.com/",$contents1)[1])[0];
     $i++;
+    if($arr[$i]['name']!=''){
     $ret = $kv->delete('icodropsproducts:'.$i);
         $kv->add('icodropsproducts:'.$i, json_encode($arr[$i],true));
     $sql='insert into ico_Analysis (name,Github_url,DataSource) values("'.$arr[$i]['name'].'","'.$arr[$i]['githuburl'].'","icodrops");';
     MySQLRunSQL($sql);
      $kv->get('icodropsproducts:'.$i);
+    }
 }
 $ret = $kv->delete('icodropsproducts:all');
         $kv->add('icodropsproducts:all', json_encode($arr,true));
